reasoning/uncertainty-quantification/uq_widget/requirements_uqwidget.txt
colour==0.1.5
dash==2.9.1
dash_bootstrap_components==0.13.1
dash_daq==0.5.0
jupyter_dash==0.4.2
numpy==1.24.2
pandas==1.5.3
Pillow==9.4.0
plotly==5.13.1
